# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Buy or Wait?","description":"Deterministic affordability engine over a 90-day daily cash-flow projection.","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/api/samples":{"get":{"summary":"List Samples","description":"The 25 solved cases, for the frontend to load as starting points.","operationId":"list_samples_api_samples_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response List Samples Api Samples Get"}}}}}}},"/api/samples/{request_id}":{"get":{"summary":"Get Sample","description":"Rebuild a full analyze payload from the dataset, so the frontend can\nload a realistic case and then let the user edit it.","operationId":"get_sample_api_samples__request_id__get","parameters":[{"name":"request_id","in":"path","required":true,"schema":{"type":"string","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeRequest-Output"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyze":{"post":{"summary":"Analyze Request","operationId":"analyze_request_api_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeRequest-Input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalyzeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analyze/batch":{"post":{"summary":"Analyze Batch","description":"Sequential rather than concurrent: the engine is pure CPU work with no\nI/O to overlap, so threads would add contention without throughput.","operationId":"analyze_batch_api_analyze_batch_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AnalyzeRequest-Input"},"type":"array","title":"Payloads"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AnalyzeResponse"},"type":"array","title":"Response Analyze Batch Api Analyze Batch Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/accuracy":{"get":{"summary":"Accuracy","description":"Score the engine against the 25 solved samples.\n\nRuns the real pipeline, not a cached result file: if the engine regresses,\nthis endpoint shows it immediately.","operationId":"accuracy_api_accuracy_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccuracySummary"}}}}}}}},"components":{"schemas":{"AccuracyRow":{"properties":{"request_id":{"type":"string","title":"Request Id"},"user_id":{"type":"string","title":"User Id"},"request_type":{"type":"string","title":"Request Type"},"currency":{"type":"string","title":"Currency"},"requested_amount":{"type":"number","title":"Requested Amount"},"expected_status":{"type":"string","title":"Expected Status"},"predicted_status":{"type":"string","title":"Predicted Status"},"status_match":{"type":"boolean","title":"Status Match"},"expected_method":{"type":"string","title":"Expected Method"},"predicted_method":{"type":"string","title":"Predicted Method"},"method_match":{"type":"boolean","title":"Method Match"},"expected_amount":{"type":"number","title":"Expected Amount"},"predicted_amount":{"type":"number","title":"Predicted Amount"},"relative_error":{"type":"number","title":"Relative Error"},"rejections":{"items":{"type":"string"},"type":"array","title":"Rejections"}},"type":"object","required":["request_id","user_id","request_type","currency","requested_amount","expected_status","predicted_status","status_match","expected_method","predicted_method","method_match","expected_amount","predicted_amount","relative_error","rejections"],"title":"AccuracyRow"},"AccuracySummary":{"properties":{"total":{"type":"integer","title":"Total"},"status_correct":{"type":"integer","title":"Status Correct"},"method_correct":{"type":"integer","title":"Method Correct"},"amount_within_2pct":{"type":"integer","title":"Amount Within 2Pct"},"amount_within_10pct":{"type":"integer","title":"Amount Within 10Pct"},"median_relative_error":{"type":"number","title":"Median Relative Error"},"rows":{"items":{"$ref":"#/components/schemas/AccuracyRow"},"type":"array","title":"Rows"}},"type":"object","required":["total","status_correct","method_correct","amount_within_2pct","amount_within_10pct","median_relative_error","rows"],"title":"AccuracySummary"},"AnalyzeRequest-Input":{"properties":{"profile":{"$ref":"#/components/schemas/ProfileIn-Input"},"events":{"items":{"$ref":"#/components/schemas/EventIn-Input"},"type":"array","title":"Events"},"payment_options":{"items":{"$ref":"#/components/schemas/PaymentOptionIn-Input"},"type":"array","title":"Payment Options"},"request":{"$ref":"#/components/schemas/RequestIn-Input"},"resolution":{"type":"string","title":"Resolution","default":"daily"}},"type":"object","required":["profile","events","request"],"title":"AnalyzeRequest"},"AnalyzeRequest-Output":{"properties":{"profile":{"$ref":"#/components/schemas/ProfileIn-Output"},"events":{"items":{"$ref":"#/components/schemas/EventIn-Output"},"type":"array","title":"Events"},"payment_options":{"items":{"$ref":"#/components/schemas/PaymentOptionIn-Output"},"type":"array","title":"Payment Options"},"request":{"$ref":"#/components/schemas/RequestIn-Output"},"resolution":{"type":"string","title":"Resolution","default":"daily"}},"type":"object","required":["profile","events","request"],"title":"AnalyzeRequest"},"AnalyzeResponse":{"properties":{"decision":{"$ref":"#/components/schemas/DecisionOut"},"projection":{"$ref":"#/components/schemas/ProjectionOut"},"reasoning":{"$ref":"#/components/schemas/ReasoningOut"}},"type":"object","required":["decision","projection","reasoning"],"title":"AnalyzeResponse"},"DecisionOut":{"properties":{"amount_safe_to_pay":{"type":"number","title":"Amount Safe To Pay"},"affordability_status":{"type":"string","title":"Affordability Status"},"recommended_payment_method":{"type":"string","title":"Recommended Payment Method"},"payment_plan":{"items":{"$ref":"#/components/schemas/PaymentOut"},"type":"array","title":"Payment Plan"},"earliest_date_for_full_payment":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Earliest Date For Full Payment"},"spending_changes_needed":{"items":{"type":"string"},"type":"array","title":"Spending Changes Needed"},"decision_explanation":{"type":"string","title":"Decision Explanation"}},"type":"object","required":["amount_safe_to_pay","affordability_status","recommended_payment_method","payment_plan","earliest_date_for_full_payment","spending_changes_needed","decision_explanation"],"title":"DecisionOut"},"EventIn-Input":{"properties":{"event_id":{"type":"string","title":"Event Id"},"event_type":{"type":"string","title":"Event Type"},"description":{"type":"string","title":"Description","default":""},"category":{"type":"string","title":"Category","default":""},"direction":{"type":"string","title":"Direction"},"amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Amount"},"currency":{"type":"string","title":"Currency"},"event_date":{"type":"string","format":"date","title":"Event Date"},"settlement_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Settlement Date"},"status":{"type":"string","title":"Status","default":"settled"},"flexibility":{"type":"string","title":"Flexibility","default":"fixed"},"minimum_allowed_amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Minimum Allowed Amount"},"linked_event_id":{"type":"string","title":"Linked Event Id","default":""}},"type":"object","required":["event_id","event_type","direction","currency","event_date"],"title":"EventIn"},"EventIn-Output":{"properties":{"event_id":{"type":"string","title":"Event Id"},"event_type":{"type":"string","title":"Event Type"},"description":{"type":"string","title":"Description","default":""},"category":{"type":"string","title":"Category","default":""},"direction":{"type":"string","title":"Direction"},"amount":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Amount"},"currency":{"type":"string","title":"Currency"},"event_date":{"type":"string","format":"date","title":"Event Date"},"settlement_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Settlement Date"},"status":{"type":"string","title":"Status","default":"settled"},"flexibility":{"type":"string","title":"Flexibility","default":"fixed"},"minimum_allowed_amount":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Minimum Allowed Amount"},"linked_event_id":{"type":"string","title":"Linked Event Id","default":""}},"type":"object","required":["event_id","event_type","direction","currency","event_date"],"title":"EventIn"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PaymentOptionIn-Input":{"properties":{"payment_option_id":{"type":"string","title":"Payment Option Id"},"payment_method":{"type":"string","title":"Payment Method"},"payment_amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Payment Amount"},"number_of_payments":{"type":"integer","title":"Number Of Payments","default":1},"first_payment_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"First Payment Date"},"payment_frequency_days":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Payment Frequency Days"},"financing_fee":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Financing Fee","default":"0"},"total_payable_amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Total Payable Amount"}},"type":"object","required":["payment_option_id","payment_method","payment_amount"],"title":"PaymentOptionIn"},"PaymentOptionIn-Output":{"properties":{"payment_option_id":{"type":"string","title":"Payment Option Id"},"payment_method":{"type":"string","title":"Payment Method"},"payment_amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Payment Amount"},"number_of_payments":{"type":"integer","title":"Number Of Payments","default":1},"first_payment_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"First Payment Date"},"payment_frequency_days":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Payment Frequency Days"},"financing_fee":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Financing Fee","default":"0"},"total_payable_amount":{"anyOf":[{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"},{"type":"null"}],"title":"Total Payable Amount"}},"type":"object","required":["payment_option_id","payment_method","payment_amount"],"title":"PaymentOptionIn"},"PaymentOut":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"amount":{"type":"number","title":"Amount"}},"type":"object","required":["date","amount"],"title":"PaymentOut"},"ProfileIn-Input":{"properties":{"user_id":{"type":"string","title":"User Id","default":"user"},"home_currency":{"type":"string","title":"Home Currency"},"current_available_balance":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Current Available Balance"},"minimum_balance_to_keep":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Minimum Balance To Keep"},"financial_priorities":{"items":{"type":"string"},"type":"array","title":"Financial Priorities"},"expense_categories_to_protect":{"items":{"type":"string"},"type":"array","title":"Expense Categories To Protect"},"expense_categories_user_is_willing_to_reduce":{"items":{"type":"string"},"type":"array","title":"Expense Categories User Is Willing To Reduce"},"expense_categories_user_is_willing_to_stop":{"items":{"type":"string"},"type":"array","title":"Expense Categories User Is Willing To Stop"},"payment_methods_user_will_consider":{"items":{"type":"string"},"type":"array","title":"Payment Methods User Will Consider"},"max_installment_months":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Installment Months"}},"type":"object","required":["home_currency","current_available_balance","minimum_balance_to_keep"],"title":"ProfileIn"},"ProfileIn-Output":{"properties":{"user_id":{"type":"string","title":"User Id","default":"user"},"home_currency":{"type":"string","title":"Home Currency"},"current_available_balance":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Current Available Balance"},"minimum_balance_to_keep":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Minimum Balance To Keep"},"financial_priorities":{"items":{"type":"string"},"type":"array","title":"Financial Priorities"},"expense_categories_to_protect":{"items":{"type":"string"},"type":"array","title":"Expense Categories To Protect"},"expense_categories_user_is_willing_to_reduce":{"items":{"type":"string"},"type":"array","title":"Expense Categories User Is Willing To Reduce"},"expense_categories_user_is_willing_to_stop":{"items":{"type":"string"},"type":"array","title":"Expense Categories User Is Willing To Stop"},"payment_methods_user_will_consider":{"items":{"type":"string"},"type":"array","title":"Payment Methods User Will Consider"},"max_installment_months":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Installment Months"}},"type":"object","required":["home_currency","current_available_balance","minimum_balance_to_keep"],"title":"ProfileIn"},"ProjectionOut":{"properties":{"points":{"items":{"$ref":"#/components/schemas/ProjectionPoint"},"type":"array","title":"Points"},"minimum_balance":{"type":"number","title":"Minimum Balance"},"trough_date":{"type":"string","format":"date","title":"Trough Date"},"trough_balance":{"type":"number","title":"Trough Balance"},"headroom":{"type":"number","title":"Headroom"}},"type":"object","required":["points","minimum_balance","trough_date","trough_balance","headroom"],"title":"ProjectionOut"},"ProjectionPoint":{"properties":{"date":{"type":"string","format":"date","title":"Date"},"balance":{"type":"number","title":"Balance"}},"type":"object","required":["date","balance"],"title":"ProjectionPoint"},"ReasoningOut":{"properties":{"detected_series":{"items":{"$ref":"#/components/schemas/SeriesOut"},"type":"array","title":"Detected Series"},"confirmed_future_events":{"items":{"$ref":"#/components/schemas/EventIn-Output"},"type":"array","title":"Confirmed Future Events"},"rejected_plans":{"items":{"type":"string"},"type":"array","title":"Rejected Plans"},"income_projected":{"type":"boolean","title":"Income Projected"}},"type":"object","required":["detected_series","confirmed_future_events","rejected_plans","income_projected"],"title":"ReasoningOut"},"RequestIn-Input":{"properties":{"request_id":{"type":"string","title":"Request Id","default":"request"},"request_date":{"type":"string","format":"date","title":"Request Date"},"request_type":{"type":"string","title":"Request Type","default":"purchase"},"requested_amount":{"anyOf":[{"type":"number"},{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"}],"title":"Requested Amount"},"desired_completion_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Desired Completion Date"},"allows_partial_payment":{"type":"boolean","title":"Allows Partial Payment","default":false},"request_text":{"type":"string","title":"Request Text","default":""}},"type":"object","required":["request_date","requested_amount"],"title":"RequestIn"},"RequestIn-Output":{"properties":{"request_id":{"type":"string","title":"Request Id","default":"request"},"request_date":{"type":"string","format":"date","title":"Request Date"},"request_type":{"type":"string","title":"Request Type","default":"purchase"},"requested_amount":{"type":"string","pattern":"^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$","title":"Requested Amount"},"desired_completion_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Desired Completion Date"},"allows_partial_payment":{"type":"boolean","title":"Allows Partial Payment","default":false},"request_text":{"type":"string","title":"Request Text","default":""}},"type":"object","required":["request_date","requested_amount"],"title":"RequestIn"},"SeriesOut":{"properties":{"category":{"type":"string","title":"Category"},"direction":{"type":"string","title":"Direction"},"period_days":{"type":"integer","title":"Period Days"},"amount":{"type":"number","title":"Amount"},"monthly_equivalent":{"type":"number","title":"Monthly Equivalent"},"last_seen":{"type":"string","format":"date","title":"Last Seen"},"flexibility":{"type":"string","title":"Flexibility"},"occurrences":{"type":"integer","title":"Occurrences"}},"type":"object","required":["category","direction","period_days","amount","monthly_equivalent","last_seen","flexibility","occurrences"],"title":"SeriesOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}